server: catch exceptions when calling unit Ctor
A common function called in ugen constructor functions is RTAlloc, the real-time memory allocation function. If RTAlloc fails (often by requesting too much memory), it throws an exception, but this exception is never caught in SC. If the ugen doesn't check for a failed RTAlloc itself, it may instantly crash the server.

This commit causes scsynth to catch all exceptions thrown during the Ctor function. If an exception is caught, the failure is graceful: an error message is printed to the user and the misbehaving ugen is set to output 0.

This does not solve all potential RTAlloc crashes, but it catches many of them.