class library: use recorder for recording patterns
This patch unifies the recording functionality. It makes in more uniform in terms of file names and protects it better from failure.

This patch adds one argument to pattern:record, namely the number of channels for the monitoring. This allows us e.g. to listen to a stereo mix down of a multichannel pattern while recording it on the full number of channels.