scsynth: CoreAudio: workaround for rate check
The scsynth CoreAudio driver checks to make sure that the sample rates of the input and output devices are the same, and if they are not, an error message is produced and scsynth refuses to boot. This commit skips the check if the number of input bus channels is 0. This allows the command-line option "-i 0" to serve as a simple workaround if the user would prefer to simply ignore the input device.