class lib: revise "cleaning up temp synthdefs"
As part of an effort to clean up initial post window output, this commit causes the "Cleaning up temp synthdefs..." post window output to be suppressed if there are no temporary SynthDefs to clean up. For most users, this gets rid of the message entirely, since most SynthDefs today are done in memory.

It also revises the message to "Cleaning up temporary SynthDefs...".