Update Lexer to allow symbols like \1a
Previously, symbols beginning with a backslash and combining digits and letters would throw a parse error if the first character was a digit, even though the same symbol could be notated without a problem in single quotes. This commit gets rid of some unnessary goto splits in the lexer that were causing the issue.