Archive test output to save space.
The output of these tests is unfortunately but (in my opinion)
necessarily massive (about 41 MB). Luckily, the output files
compress very well and this new archive is about 6M. This shouldn't
have a major impact on running the tests since they should be run
manually and in special circumstances anyway.