test_script.scd: add some lines to give easy access to tests
I have observed some long sequences of tests crashing the interpreter
for an unknown reason. Since output seems to be consistent and
since running the tests in isolation seems to resolve the problem,
I am adding a few lines in the main script file that give direct
access to the tests that are left hanging (in my situation).