scsynth: initialize memory for SF_INFO structs
This fixes an uninitialized memory issue in /b_read and /b_readChannel, which was causing nondeterministic failures in Buffer.cueSoundFile. libsndfile's sf_open takes as input a pointer to an SF_INFO struct, to which it writes its soundfile metadata. However, the SF_INFO struct's "format" parameter of the struct must be set to zero, otherwise (as best as I can tell from the sf_open docs) the audio is read in a RAW format. Failure to initialize the SF_INFO memory is therefore a bad idea, and zeroing out the memory ensures that the RAW format is not accidentally invoked.