class library: fix possible sync problem in NodeProxy clear
`MixedBundle` allows to schedule a function when sending happens. This
is what was used to schedule into the future the clearing of the old
group. It is much better to just delegate this to the server. Because
group  objects are registered by NodeProxy with the NodeWatcher, it is
also released correctly. ( `NodeWatcher.newFrom(s).nodes`).

We need to add a very small offset between the bundle timestamps, so their order is guaranteed on the server.