class library: NodeProxy - tolerant handling of nested output
In many situations, one would expect `proxy.ar` to return a single
channel, but it returns an array (there are good reasons for this,
discussed in NodeProxy help under routing).

But in order to avoid being overly strict, we treat a singleton channel
as a channel for a proxy _input_ (that is the result of the UGen
function).

It will still warn if there are more UGens in the nested arrays.