lang: Protect against Bus overwrites where possible
When a Bus object is provided as the argument for an out, we have an opportunity to validate that we're outputting a reasonable number of channels - we should do that. Outputting more than the size of your bus transparently overflows into whatever bus is next numerically, which can cause hard to diagnose bugs.