ScIDE: docManager: properly handle setText request
funcID is nil in many instances, which would cause
handleSetDocTextScRequest to bail out early while trying to parse a
null node as a string.

After this commit, funcID is only parsed if it is not null, and
a function call to Document.executeAsyncResponse is, again, only
made if the funcID is non-null.