class library: adverb argument for reduce
The method `reduce` takes a binary operator selector (or an appropriate
function) and applies it to the collection’s items incrementally.

This patch allows us to pass an adverb to the operator, as commonly
expected for binary operations. If the operator is a function, the
adverb is the third argument.

These two expressions return the same value:
```
[1, 2] *.t [10, 20, 30]
[[1, 2], [10, 20, 30]].reduce('*', 't')
```

This feature is justified because it exposes the interface of binary
operators and calculations by reduce are very compact.