scdoc: Append "| SuperCollider Help" to all <title> elements
This commit appends the suffix "| SuperCollider [version] Help" to the <title> element of every generated HTML file in SCDoc. (It's not possible to dynamically get the version in some static files like Search and Browse, so I just added "| SuperCollider Help" to those.)

Also, added HTML5 doctype and lang attributes while we're at it.