updating scdoc css towards #2943
updating scdoc css towards #2943
this PR does basically consist of changing the css font for scdoc. Century Gothic is basically one of the most used family fonts in frontend development and overall UI/UX design. sometimes the simple act of changing a font style to century gothic, produces an improvement of the overall Looking of an HTML page. that's i would kindly like to suggest that we update the scdoc css family font to css in accordance with the telephon suggestion on 2943. i suggest that we do the same on the other css documents. but i will wait for feedback on this PR on this regard, before moving into opening another PR, with the same ideas.
Looking forward
Kind regards 
T. (Tiago)