scdoc: redesign subheader
This commit improves the design of the subheader. It is no longer left-aligned. The source filename is reduced to include only the file name (which can still be viewed in full by hovering over or clicking), the superclasses are incorporated into the title. The "see more..." line for subclasses is cut down from 13 classes to 4.