class library: plot relative to target synth
Sometimes, one needs to plot values from a synth that depends on node
order. E.g. a synth that reads from a bus. This patch allows us to
determine the point at which the plotting synth is inserted in the node
order. It is inserted _after_ the target node. By default, this is the
default group of the default server. This follows the `asTarget`
interface, as documented in scdoc.