Supernova: use LIB_SUFFIX for plugin loading
Commit 547a18c introduced the CMake variable LIB_SUFFIX to determine the
system library folder, and made supernova load plugins in both /usr/lib
and /usr/lib64.
On some systems (eg archlinux64) lib64 is symlinked to lib, supernova then
tries to load the plugins twice and crashes.
We want supernova to load only the folder in which the plugins have been
installed.