ide: Fix most code execution issue with webview
Note that shift+enter for line execution doesn't work - this is unlikely to ever work, as the shift+enter is always eaten by the browser no matter what (even if intercepted and explicitly ignored by JS, the event never makes it back to the application). However, if "execute line" is set to something other than shift+enter, it works correctly.