scdoc: Fix rippling links on hover
This commit fixes links in SCDoc, particularly those in the top drop-
down menus, changing size when the cursor hovers over them. This is due
to a "padding-bottom: 1px" line in the CSS that attempts to compensate
for "text-decoration: underline" not appearing when the font size is
less than 11pt, but I don't see any such problem.