class library: implement correct stop behaviour for Task
Routines can’t stop in between yields. Nevertheless, `Routine`
implements a trick in `stop`: it checks if it is in the current thread
and if it is, it will yield nil, breaking calculation right where it is.

Because a `Task` has a routine and is not a routine, it needs to do
this check as well, but this time for its stream.