UnitTest: add option to hide details of passed assertions
When assertions succeed, we don't usually need to see the details;
they just clutter up the output.  So extend #-assert's API with an
extra details parameter, to optionally allow separation of the main
test description message from the details, and introduce a new
class variable passVerbosity, which if set to UnitTest.brief,
suppresses output of the details whenever an assertion passes.

The responsibility of formatting the report of an individual test
result should either fall entirely with UnitTestResult, or not all,
but not partially.  The former seems like a more sensible choice, so
pass the new details parameter to UnitTestResult in order that it may
handle the formatting of those details along with the other info.

This should be 100% backwards-compatible for existing callers.

Partially addresses #3365.