class library: only schedule messages if necessary
The message schedBundleArrayOnClock is sent to a number in order to
send off an OSC bundle a after certain time. If that time equals zero,
there is no need to add the task to the clock’s queue. This commit adds
a case where the bundle is send right now without scheduling it on the
clock in that case.