Remove misleading deprecation of old support directory
On UNIX-like systems, if SuperCollider is compiled from source into
the user's home directory via:

    cmake -DCMAKE_INSTALL_PREFIX=$HOME
    make install

then Platform.systemAppSupportDir and Platform.resourceDir will be set
to

    $HOME/share/SuperCollider

This is due to CMakeLists.txt containing the definition of SC_DATA_DIR:

    elseif(UNIX)
            add_definitions(-DSC_DATA_DIR="${CMAKE_INSTALL_PREFIX}/share/SuperCollider")
    endif()

which is then consumed by common/SC_Filesystem_unix.cpp.

Unfortunately, this resulting path is identical to the old location
for the *user* support directory, so it triggers a misleading warning.
However this deprecation was introduced by 2b77ef67 in August 2011,
so it should be safe to remove it by now.

Fixes #3367.