Update to non-expiring slack invite
I see that there was recently (#3647) an update to the slack invite link. Perhaps changing to this link (that's been around for [8 months or so](http://new-supercollider-mailing-lists-forums-use-these.2681727.n2.nabble.com/switch-to-Slack-td7634241.html)) would help the churn. Downside is it is a 3rd party service, not an official Slack page.

cc @brianlheim