class library: remove pushNow argument.
This can be a very confusing option when set to false. It creates a
document which will be linked properly to the environment only after
losing focus and then coming back. For documents that are not in focus,
on the other hand, it is much better not to push.

This PR is an API change.
It avoids user errors.