class library:  fix document with linked environment
Instead of relying on user code, test if the document is in front, and
push the environment correctly if this is the case. Save the current
environment only if we are not already in a document that has a saved
environment.

We add a getter for `savedEnvir` for introspection.

#testing: Test all methods that rely on setting the environment of a
document.