class library: remove pushNow argument.
This argument (`pushNow`) can be a very confusing option when set to
false. It creates a document which will be linked properly to the
environment only after losing focus and then coming back. For documents
that are not in focus, on the other hand, it is much better not to
push. This avoids user errors.

This is an API change.

#testing: Test calls of linkDoc and unlinkDoc to subclasses of
EnvironmentRedirect, e.g. ProxySpace.linkDoc. As this is an API change,
it might change the behaviour of user code (though it now behaves as
probably used in most cases, and was broken before).