class lib: Fix "nonexistent class extension" errors in non-Qt builds
This commit fixes two errors about nonexistent class extensions on non-Qt
sclang builds. These are class extensions located in the deprecated
directories that are not correctly placed in "GUI" subdirectories.

This also relocates WiiMoteGUI into a GUI subdirectory. There's not much
impact since WiiMoteGUI is about to be removed and it wasn't causing an
error, but hey, might as well.

Fixes #3769.