sclang: fixes crash on call to Array::lace with empty subarrays
Regarding issue #3716, this fixes two bugs:

a) Calls to lace with no length argument would fail if any elements in the
array weren't subarrays. Now lace will assume the max length of each
subelement is one.

b) Calls to lace with empty subarrays would crash due to a divide-by-zero
error in running modulo to calculate the lace index on a subarray of length
zero. Now the code simply copies the empty subarray into place in the ouput
array.