class library: NodeProxy crossfade shape
Setting the gate of an envelope to a negative value makes a forced
release. What remained unnoticed here was that the forced release shape
is always linear, regardless of the envelope’s shape. This commit
replaces the forced release by a normal release. The fadeTime is set,
so forcing wasn’t necessary.

This fixes #3775.

testing:
- test how node object replacement and freeing sounds.
- this may change the behaviour of SynthDefs that are separately made
and which do not define a `fadeTime` parameter to specify the release
time. A subsequent commit will try to minimize such impact.