sclang: narrow scope of setting nil in Array::lace to empty arrays
Regarding issue #3716, based on feedback on the PR, we will only
set nil in the target array if trying to copy a zero-length array.
Note that this leaves the code behaving as the unmodified version
of the code, which is that objects in the callee array that are
neither lists or arrays will not be copied.