class library: function play fade envelope when necessary
Currently, the GraphBuilder that wraps the UGens in function.play
always makes a fade envelope, unless the output is a scalar.

Instead, this should be done only when absolutely necessary. When the
Ugen graph has already an envelope that is able to release the synth,
we avoid making another envelope.

This fixes #3768.

#testing: this may have an impact on some past workarounds tha uses
{ }.play, or internally makes use of the GraphBuilder. This is an API
change.