sclang: fixes hardlock on floating point increments to Integer::forBy
Previously floating point arguments to the special optimizations on the
Integer::forBy function would be cast to Integer, which meant that if they
were relying on fractional increments those could be lost in the cast,
resulting in a nonterminating loop. This commit adds support for floating
point arguments. If any of the arguments are floating point they are all
cast to floating point and the loop proceeds using floating point.