sclang: fix hang on floating point arguments to Integer:forBy
Previously Integer:forBy would truncate floating point arguments, to integers,
which could potentially result in an infinite loop by the Integer:forBy
optimized special operation. This modification checks for any of the arguments
to be floating point, and if detecting that casts all of the arguments to
floating point. It also adds appropriate testing coverage.