lang: Fix remaing users of QObjectProxy
QObjectProxy was used in a few places as a sort of psuedo-generic type, probably just to avoid having to create a type converter for a new type. We now have the right type conversions for these, and in general this is a bad pattern: we should stick to (roughly) 1:1 relationships between sclang types and C++ types, else it becomes very hard to map between them correctly.