sclang: Simplify TreeView signals
TreeView redirected it's signals in a non-straightfoward way so it could receive and send QcTreeWidget::ItemPtr (which represent individual items in the tree). But, signals can cast types, so all we need here is (a) a cast operator for our ItemPtr, and (b) forwarding for all the signals on the base class. Rather than changing names of singals as before, I just kept them consistant with the base class and updated the TreeView.sc to match. We get a few more for free that no one bothered to expose previously.