class library: function plot uses loadToFloatArray
Because getToFloatArray is broken in some specific boundary case, for
now we just use loadToFloatArray instead. This used to be the case in
previous versions of the method, so this change doesn’t break expected
functionality.

Note that for now this again limits plotting to local servers.

By curing the symptom, this fixes #3873.

getToFloatArray still needs to be fixed.