Disable some SerialPort tests
This commit disables one non-functioning SerialPort test, and skips
SerialPort tests if the OS is Windows or it is determined that the
socat utility is not installed.

There is currently no skip mechanism in UnitTest, and due to time
constraints we can't implement it right now. So I've just hacked it
together using a guard clause.