help-browser: use webchannel to submit eval requests
shift-enter needs to be intercepted at the JS level and call a slot on a C++ object
in order to evaluate a line. This shortcut isn't propagated normally.
This commit adds QWebChannel and QWebSockets to the dependencies
and uses them to allow JS to request the help browser C++ code to evaluate a line.

Also some changes to JS code and SCDoc renderer code

Note that the WebChannel should only be loaded on pages with code to
execute.

The WebChannel is constructed in the IDE's main() and then a new wrapper
for IDE objects is published to it. On the JS side, evaluateLine() is
called whenever shift-enter on a code snippet is detected.