[scide] Adds some refinement based on testing on MacOS X.
Testing all default provided color themes on MacOS exposed a few issues:

Light text on dark backgrounds breaks buttons on OS X which are always rendered
with the system background color of white, making the text unreadable. We
clamp colors to a certain lightness on OS X to prevent this.

Generally the active tab needs to be highlighted with a different color.

There are still a few questions and refinements here, going to test back on
my Linux machine later this evening. Thinking of swapping the active tab
and inactive tab colors.