sclang: Defer showing of dialog
If we show dialog while we're in the middle of an Interpret call (as can be the case now), and the idle loop of the dialog tries to process SuperCollider timer events, we hang (since we're already mid-interpret). We need to launch the dialog without SC on the stack, so we defer. QTimer doesn't actually defer things with time==0, so we have to do 1ms.