[scide] adjustments to ui palette value separation heuristic
The old heuristic reacted to a value difference of only 16, but this allowed
for some cases where the user could provide colors with input values that would
not be separated by 32 units of value, which from testing makes the UI less
usable. The new code always distances the colors by 32 units of value.