Plotter: ensure resampled values always start and end on the correct values.
Previous used resamp1, which allows floating-point resample argument,
resulting in a last value which isn’t the last in the data, but rather
one just short of it. For example resampled data that’s fairly linear
may appear to not reach its final value. This method ensures correct
start and end values, and has evenly spaced sampling within those
values that is a least ‘resolution’ pixels.