travis macos: don't update brew, fix build error [skip appveyor]
latest version of travis seems to be barfing when trying to install
ccache. dunno why, but either way forcing homebrew not to update seems
to fix it. we intended to do this anyway but mistyped the environment
variable names! this shaves quite a bit of time off the build.

also cat the un-xcpretty'd build log when there's a failure, which in this case would
have helped to diagnose the problem more quickly.