scide: Fix weird colors on legacy theme loading
This commit fixes a bug where old themes from pre-3.7 versions of the IDE would load with weird colors. This is because the text color and background fields defaulted to black when not specified. This fixes theme deserialization so that text defaults to black and backgrounds default to either black or white, whichever contrasts better.

For #4218.