class library: fix setbus_control Event type
The Event type \setbus_control is intended to be used for crossfading
between values on a control bus. There are cases in which a previously
playing synth still writes on the bus after the fadeTime (this can
happen due to a slightly later release of the other crossfade
envelope). Then the target value that the control bus should have after
transition is lost.

Adding one control period to the envelope as a hold value is a way to
avoid this glitch.

This fixes  #4295.