class library: node proxy - fix queuing problem
This makes sure that:
a) SynthDefs have a unique name instead of a channel id
b) SynthDefs are freed even when the bundle is canceled

Not setting the nodeID to nil is ok, nothing substantial depends on it.

For some reason, `addPrepare` needs to be replaced by a direct adding.
I would have thought that either way should be OK, but `addPrepare`
will cause SynthDef not found errors under load (when bundles are
canceled). This may need further checking, including the server
behaviour.