scsynth: Set default socket address to 127.0.0.1
This commit changes the default socket address, user-settable from the command line via the -B option, to 127.0.0.1 rather than 0.0.0.0. For most users running laptops connected to wifi, 0.0.0.0 is a dangerous default, since it allows anyone on the local network to send OSC messages to scsynth.

See issue #4496 for more explanation.