scide: Fix a dangling pointer: Document::mLastActiveEditor
When a GenericCodeEditor is destroyed but a Document refers to it via
mLastActiveEditor, said pointer is dangling and can cause a crash. For
example, creating a split, removing all splits, and then recompiling the
class library can cause a crash.

This commit is a quick bandage, simply setting mLastActiveEditor to a
null pointer in the GenericCodeEditor's destructor. A more thorough
solution would be a stack of the most recently active editors, but the
point is to just prevent an IDE crash right now.