class library: error in parallel streams shouldn't affect main player
A TaskProxy can return several independent streams, any of which may
throw an error. They shouldn’t stop the main (internal) player, but
just the player of that stream. In the case of a single internal
player, this is identical (play calls playOnce and keeps the result as
internal player).