class library: fix def proxy dup
When no key was passed into `copy` for an Ndef (and analogous classes),
it would fail obscurely. This happens easily, e.g. in `dup`. It is good
to throw an error, because usually one expects copy to return a
separate object. But in cases like `dup`, a copy isn’t needed.

This fixes #4685.