Fix some instances of broken spacing
There seems to be a corner-case in PROSE handling which leads to missing
spaces.  In particular, if there is a code::word::
followed by normal text on a new line, HTML omits a space
between the closing tag and the text of the new line.
I can't tell how other browsers react, but here is how Lynx currently
renders Classes/EnvGen. My in-development Emacs SCDoc
renderer exposes the same problem.
Note the "word" timeScaleare
Also note the extra spaces in "levelScale , levelBias ,"

Description

   Plays back break point envelopes. The envelopes are instances of the
   Env class. The envelope and the arguments for levelScale , levelBias ,
   and timeScaleare polled when the EnvGen is triggered and remain
   constant for the duration of the envelope.