Versioning reform
this commit updates how versioning is communicated across the codebase.
the most important change is that SCVersion.txt now separates the
'patch' and 'tweak' version components in new CMake variables called
SC_VERSION_[MAJOR,MINOR,PATCH,TWEAK]. 'patch' is always numeric, while
'tweak' is the optional component used for '-beta1', '-rc2'. the old
variables are kept for backwards compatibility.

we use these to set the VERSION in the CMake call to project(), which
fixes #4648. however, this requires that we strip the 'tweak' component
since only digits and periods are allowed in version strings. so, we
replace references to PROJECT_VERSION with SC_VERSION where appropriate.

this also allows us to be more precise with versioning in the class
library. previously, we only had a method that would return the
".4-beta1" part of the version string, Main.scVersionPostfix. this
commit deprecates that method and replaces it with Main.scVersionPatch
and Main.scVersionTweak. now, we can also update Main.versionAtMost and
Main.versionAtLeast to check the patch version as well.

scsynth and supernova's `/version.reply` contents are kept the same for
backwards compatibility.

documentation added/updated for all affected methods.

tests added for Main.versionAtMost/Least

testing checklist:
- try re-configuring cmake multiple times to confirm #4648 is fixed
- check that outputs of Main.version, Main.scVersionFoo commands are
unchanged
- check that scsynth and supernova replies to /version are unchanged