class library: fix quant in Pdef for short patterns
Where the old source pattern is shorter than the quant value, Pdef /
EventPatternProxy would start the new pattern too early. This is
because Pfindur only cuts and doesn’t extend. This change checks if the
stream has ended and adds a silent event that holds the time until the
new pattern comes in.