CMakeLists.txt: Adding option to build with system Ableton Link.
Adding SYSTEM_ABLETON_LINK (defaults to OFF) option to indicate, that the system version of Ableton Link should be used.

lang/CMakeLists.txt: Calling find_package() for AbletonLink with appropriate NAMES.

Currently downstreams package AbletonLink as ableton-link and link, which places AbletonLinkConfig.cmake in the wrong paths
(e.g. /usr/share/ableton-link and /usr/lib/cmake/link for Debian and Arch respectively).
Therefore find_package() for AbletonLink has to be called with appropriate NAMES, so that both the path *and* the file can be resolved.
Also, setting REQUIRED, so that cmake will fail generically on not finding AbletonLink.