Adding script to create a source tarball.
package/create_source_tarball.sh:

This script accepts two parameters:
* v: The version (aka. the tag) the source tarball is supposed to be
  created for (mandatory).

* s: The signing entity (an e-mail address or key ID) to sign the source
  tarball with a PGP private key.

The script creates a tarball of the form
SuperCollider-X.Y.Z-Source.tar.bz2, which extracts to a directory of the
form SuperCollider-X.Y.Z-Source/ (here X.Y.Z denote a version without
the conventional Version- prefix, used for the tags).
Optionally, this script creates a detached PGP signature (tarball name +
.asc).

If the BUILD_DIR environment variable is set it will be used as the
output_dir.
In this case the source tarball (and the optional detached signature)
are moved to BUILD_DIR/package and the output_dir is created
on-the-fly.
If BUILD_DIR is not provided the script defaults back to its own
location (the directory the script resides in) as output_dir.

The script utilizes a trap to cleanup temporary working directories
(even if unsuccessful).

.gitignore: Adding ignores for source tarball and detached signature
creation.

Signed-off-by: David Runge <dave@sleepmap.de>