Fix segfault in prMatchLangIP on Linux
When I tried to initialize a remote scsynth with

  Server.remote(\salt, NetAddr("192.168.7.2", 57110), ServerOptions.new.maxLogins = 4);

sclang answered by crashing with a segmentation fault.

gdb revealed that the crash happens in prMatchLangIP.  And this got me
thinking: "Maybe I should stop my VPN?"  And indeed, that fixed the
symptom.

Reading getifaddrs(3) explains it:

  The ifa_addr field points to a structure containing the interface address.
  (The sa_family subfield should be consulted to determine the format of the
  address structure.)  This field may contain a null pointer.

So fix prMatchLangIP by checking that ifa->ifa_addr is not NULL.