class library: Add a 'nil.yield' in Server's 'wait' method.
This method is used inside a Routine instance. It waits
for a server's reply. The method defines a one shot
OSCFunc instance that resumes the running thread as a response
to a server message.

The 'nil.yield' after the OSCFunc pauses the running thread
until the OSCFunc's response resume it.