class library: Server's 'wait' method reimplemented using Condition.
This method is used inside a Routine instance. It is used to
wait for a server's reply. The method defines a one shot
OSCFunc instance that sends a 'resume' message to the running
thread.

A Condition is used to pause the running thread. When the server
replies, the OSCFunc sets Condition's 'test' property to true
and sends a 'signal' message to the Condition object.