Quick fix for Pmono (and PmonoArtic) turning everything off in the default group
Commit 83e4607cd merged in PR #4787 exposed an underlying problem
in Pmono's two-part cleanup setup, which can fire a `type: off` event
with no `id` set, turning everything off in the default group.

This is an "emergency fix" that prevents Pmono's cleanup from
firing that `off` event if `id` is still not set by the time the
cleanup is executed.