Add PulseAudio backend for SuperCollider on Linux
For a discussion of why we should have a PulseAudio backend for
SuperCollider on Linux, see https://github.com/supercollider/rfcs/pull/11.

It is important to understand that this is not meant to be a replacement
for jack at all. This is just to give an option on systems where jack is
difficult to set up. jack will continue to be the default.

This PR adds support for PulseAudio, both inputs and outputs. The device
selection is done outside SC, as per PulseAudio philosophy, where
applications output to a default sink, and get input from a default
source, and the actual connections happen on the system's setup /
PulseAudio control applications.

The changes are pretty self contained, with just one C++ file added, and
then some changes to CMakeLists and READMEs as required.

The selection of frontend is done at build time, and by default the
backend will continue to be jack. This will only try to use the
PulseAudio backend if instructed explicitly to do so, using the AUDIOAPI
cmake variable. So, for users that are currently happy with their setup,
using jack, this change will not affect them at all and will be
completely transparent.

On systems where it is difficult to use jack, the user / maintainer can
set the cmake variable to pulseaudio, and then, and only then, it will
use the PulseAudio backend.

As you can see the code is not too complex, and should be easy to
maintain. Note that the expectation is that the PulseAudio backend will
be used in simple-ish systems, so we may not want to support complex /
exotic audio configurations, and encourage people to use jack for that.