Fix block size and hardware buffer size handling
Block size and hardware buffer size are handled properly now.
The DriverSetup method is supposed to return the hardware buffer size.
Then the audio callback iterates numBufs times, which is the number of
control blocks per hardware buffer period.
Updated the data move code to take into account that there can be more
than one buffers per callback.
Tested this with various control block and hardware buffer sizes
and it is working ok.
Removed my bufferFrames variable which is redundant and confusing.
Added the calculation of latency as per the advice.