Docs: Dictionary: clarify equality checks and object initialization
Connected to Issue #4573: it should be explicitly stated if a
function checks for equality or identity. A lot of the examples
used the shortcut for initializing an event, which leads to
a different behaviour regarding checks for equality. So I decided
to also change all the examples to use a correct initilization for
creating a dictionary object.