class library: fix NodeProxy mapping numChannel derivation
This fixes #5303. The issue resulted from the fact that `ProxyNodeMap` intentionally returns instances of `ControlName` that have the mapped proxy as default values. In the case of the `<<>` operator, we are only interested in the objects' control names. It would have been enough to use the keyword argument switch `.controlNames(addNodeMap: false)`, but it is much more efficient and clear as it is solved here.