audio and osc basically (almost) working. shared-array-buffer prob
- so all pieces are in place, OSC messages seem to be correctly encoded
- audio driver start/stop now in place
- but we get these errors:

Uncaught TypeError: Failed to execute 'copyToChannel' on 'AudioBuffer': The provided ArrayBufferView value must not be shared.
    at ScriptProcessorNode.ad.proc.onaudioprocess (scsynth.js:1688)
ad.proc.onaudioprocess @ scsynth.js:1688

It seems that because of `-pthread`, the HEAP is now backed by
`SharedArrayBuffer`, and that in turn cannot be used in `copyToChannel`.

Hopefully, we can just make one additional copy. If not, we're f*cked
:-O