class library: better implementation of function flop
This refactors function.flop and makes it a little easier to understand. By the generalisation, we gain an easy way to implement inEnvir, in a way that such a function bound to an environment can be called like any other function.

For inEnvir, this comes at a performance cost at build time but it has proven worthwhile for flop already.