class library: only one server meter by default (#5908)
This fixes #5545. ServerMeter now behaves like Scope. By result of the discussions in #5549, it may make sense to rethink the design of all server windows, and potentially an easy way to handle unique windows.

A refactoring of the server views is recommended, this is just the minimal step to fix the issue at hand.

Multiple ServerMeters can be created by directly instantiating the object (like Stethoscopse for multiple scopes).